feat: pass full response object to onResponse hook
Before this change, the onResponse hook was only given the response stream to work with. This was useful if you wanted to transform the response body on a chunk by chunk basis, but did not allow inspecting the status code or headers of the response! If you wanted to serve a custom error page or similar, you couldn't get at the actual `res` object in question.

So, this changes the `onResponse` hook to instead pass the whole response instead of just the stream. Now, users can inspect the status code as well.